rtt
C => S
S => C
RTT
