+++ public/scripts/client.js
@@ -47,3 +47,6 @@ socket.on("otherPlayers", (otherPlayers) => {
 socket.on("setPlayerPosition", (socketID, x, y) => {
     players[socketID].cursor.setLastReceivedPosition(x, y);
 });
+socket.on("sendRTT", (timeSent, serverReceivedTime, clientToServerTime) => {
+    console.log("C => S: " + clientToServerTime, "S => C: " + (Date.now() - serverReceivedTime), "RTT: " + (serverReceivedTime - timeSent));
+});
